removed pytorch lstm, cleaned up the 3 models and fixed ml_predict() wonkiness
