+++ sips/macros/tfm.py
@@ -1,5 +1,5 @@
 from sips.macros import macros as m
 
 
-WRITE_TO = m.PROJ_DIR + "ml/logs/"
+WRITE_TO = m.PARENT_DIR + "logs/"
 READ_FROM = m.PARENT_DIR + "data/lines/lines/"
