+++ sips/ml/ml_dir_pred.py
@@ -7,16 +7,18 @@ from sips.macros import tfm
 from sips.h import hot
 import sips.h.helpers as h
 
-from sips.ml import tf_loaders as tfls
-from sips.ml import tf_fwd
+from sips.ml import loading as tfls
+from sips.ml import fwd
 from sips.ml import tf_utils as tfu
-from sips.ml import tf_lstm as lstm
+from sips.ml import lstm
 
 
 def train_directional_predictor(datasets, test_datasets):
     # EPOCHS = 10
     BATCH_SIZE = 1
-    BUFFER_SIZE = 100
+    BUFFER_SIZE = 1
+    
+
     x, y = tfu.get_example(datasets)
 
     model, loss_fxn, optim = tfu.classify_model_core(x.shape[-2:], tf.size(y[0]))
@@ -38,26 +40,27 @@ def train_directional_predictor(datasets, test_datasets):
         if not dataset:
             continue
         for (xtr, ytr) in dataset:
-            tl, ta, correct = tf_fwd.train_step_classify(
+            tl, ta, preds = fwd.train_step_classify(
                 model, optim, loss_fxn, xtr, ytr, train_loss, train_accuracy
             )
 
-            if correct.any():
-                print("guessed correctly")
-            else:
-                print("guessed wrong")
+            # if correct.any():
+            #     print("guessed correctly")
+            # else:
+            #     print("guessed wrong")
 
         with train_summary_writer.as_default():
 
             tf.summary.scalar("loss", tl.numpy(), step=epoch)
             tf.summary.scalar("accuracy", ta.numpy(), step=epoch)
 
+        model.reset_states()
         test_dataset = random.choice(test_datasets)
 
         if not test_dataset:
             continue
         for (xte, yte) in test_dataset:
-            tel, tea = tf_fwd.test_step(
+            tel, tea = fwd.test_step(
                 model, loss_fxn, xte, yte, test_loss, test_accuracy
             )
 
@@ -65,7 +68,7 @@ def train_directional_predictor(datasets, test_datasets):
             tf.summary.scalar("loss", tel.numpy(), step=epoch)
             tf.summary.scalar("accuracy", tea.numpy(), step=epoch)
 
-        template = "Epoch {}, Loss: {}, Accuracy: {}, Test Loss: {}, Test Accuracy: {}"
+        template = "Epoch {}, Loss: {}, Accuracy: {}, Test Loss: {}, Test Accuracy: {}, Preds: {}, Actuals: {}"
         print(
             template.format(
                 epoch + 1,
@@ -73,6 +76,8 @@ def train_directional_predictor(datasets, test_datasets):
                 train_accuracy.result() * 100,
                 test_loss.result(),
                 test_accuracy.result() * 100,
+                preds,
+                ytr
             )
         )
 
@@ -81,6 +86,7 @@ def train_directional_predictor(datasets, test_datasets):
         test_loss.reset_states()
         train_accuracy.reset_states()
         test_accuracy.reset_states()
+        model.reset_states()
 
     tf.saved_model.save(model, tfm.WRITE_TO + "directional_prediction/")
